modify: integrate auth logice into userStore and also move fields and button to view component
