+++ .lintstagedrc
@@ -1,4 +1,4 @@
 {
-  "**/*.{js,jsx,ts,tsx}": ["eslint --quiet --fix", "vitest related --run"],
+  "**/*.{js,jsx,ts,tsx}": ["vitest related --run", "eslint --quiet --fix"],
   "**/*.{json,js,ts,jsx,tsx,html}": "prettier --write --ignore-unknown"
 }
