+++ src/stores/__test__/userStore.test.ts
@@ -1,16 +1,16 @@
 import { it, describe, expect } from 'vitest'
 import { act, renderHook } from '@testing-library/react'
-import { authStore } from '../authStore'
+import { userStore } from '@stores/userStore'
 
 describe('authStore', () => {
   describe('setActionType', () => {
     it('should be "login" by default', () => {
-      const { result } = renderHook(() => authStore())
+      const { result } = renderHook(() => userStore())
       expect(result.current.actionType).toBe('login')
     })
 
     it('should be "register" when change the actionType to "register"', () => {
-      const { result } = renderHook(() => authStore())
+      const { result } = renderHook(() => userStore())
       expect(result.current.actionType).toBe('login')
 
       act(() => result.current.setActionType('register'))
