+++ src/stores/index.ts
@@ -1,2 +1,1 @@
-export { authStore } from './authStore'
 export { userStore } from './userStore'
