+++ src/stores/userStore.ts
@@ -1,13 +1,20 @@
 import { create } from 'zustand'
 
-interface UserStore {
-  isAuthenticated: boolean
-  setAuthenticated: (value: boolean) => void
+export type ActionType = 'login' | 'register'
+interface State {
+  actionType: ActionType
+}
+interface Actions {
+  setActionType: (value: ActionType) => void
   reset: () => void
 }
 
-export const userStore = create<UserStore>(set => ({
-  isAuthenticated: false,
-  setAuthenticated: value => set({ isAuthenticated: value }),
-  reset: () => set({ isAuthenticated: false })
+const initialState: State = {
+  actionType: 'login'
+}
+
+export const userStore = create<State & Actions>((set) => ({
+  ...initialState,
+  setActionType: (value: ActionType) => set({ actionType: value }),
+  reset: () => set({ ...initialState })
 }))
